Make prob_calc and word_gen functions
To improve flexibility of this tool, prob_calc and word_gen are now
functions that can be called from another program. This makes using
input data of different shapes easier as the user can reshape their data
before passing it to prob_calc. In addition, word_gen now returns the
generated words, giving the user the flexibility to do what they wish
with the generated words.